v1.0.0.0rc (#8)
This commit includes new features and optimizations, as well as updated
documentation to go with the changes. Most of the features discussed in
the *Future Directions* section of the previous version's README have
been added. To see a detailed summary of everything that has changed,
see the CHANGELOG.

Special thanks to @rphair and @TerminadaPool for helping with the
README.