added support for non-compounding with incentivized minimum payments
Fixed a bug where non-compounding loans could not have incentivized
minimum payments. The previous design required `compoundFrequency`
to be `Nothing` in order to enable non-compounding, but this would
also disable the required minimum payments. Now, a separate boolean
flag is used in the datums to toggle compounding of the interest.

`compoundFrequency` was renamed to `epochDuration` and `lastCompounding`
was renamed to `lastEpochBoundary` in order to make the names more
representative of their functionality.

The helper bash scripts were also updated to use the conway era
flags with cardano-cli.