feat: added first-class support for unsecured loans
Technically, the protocol already supported unsecured loans (i.e.,
loans not backed by collateral), but it would require calculating
the collateralizations to basically equal the minUTxOValue which is
inconvenient. Also, the prior version had no way for the borrower to
signal to lenders that they are looking for an unsecured loan.

With this new version, borrowers can leave the collateral list empty
to signal the desire for an unsecured loan. Lenders can easily filter
out these loan requests if they are not interested in making unsecured
loans.